B-150 (#11)
* prueba

* T-28:
- get charges by user id

* prueba:
- medio arreglado

* typing error

---------

+++ src/main/java/backend/plugfinder/repositories/ChargerRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ChargerRepo extends CrudRepository<ChargerModel, Integer> {
+public interface ChargerRepo extends CrudRepository<ChargerModel, Long> {
 
 
 }
